This commit removes the Clone constraint from the ProgressCallback trait, changing it to:
pub trait ProgressCallback: Send

Without this, ProgressCallback would not be "object-safe", which simply
put, means it is implicitly sized to the Clone constraint and a Boxed
version of it like Box<dyn ProgressCallback> cannot be constructed.

This change places the constraint on associated arguments for functions
which accept a progress callback.